setFh(): Create file handles with `os.O_RDWR`
Use `os.O_RDWR` instead of `os.O_CREATE|os.O_RDONLY`.

`os.O_CREATE` will create the file if it doesn't exist, but we only want
to be able to lock the file if it exists. Otherwise, locking the file
and creating it can interfere with other programs that use a temporary
file that gets renamed in order to save.

If we remove `os.O_CREATE` and leave `os.O_RDONLY`, we can't use file
locking to find out if the file is locked for writing by another
program. If that's the case, the lock would succeed as a read-only lock,
eliminating that system of detection.

In order to enable detection of other locks on the file, use `os.O_RDWR`
so that we don't create the file, but can still detect whether a lock
has been placed on it by another program.